Use mutable update to defer out-of-scope errors
Previously, we let-bound an identifier to use to carry
the erroring evidence for an out-of-scope variable. But
this failed for levity-polymorphic out-of-scope variables,
leading to a panic (#17812). The new plan is to use
a mutable update to just write the erroring expression directly
where it needs to go.

While in town, I simplified the type-checking of unbound
identifiers by a tiny bit. This happily led to the elimination
of a CtOrigin.

Close #17812.

Test case: typecheck/should_compile/T17812